requirement.txt
streamlit==0.89.0
snowflake-connector-python==2.8.2
